+++ src/lib/validation/exports.ts
@@ -1,17 +1,19 @@
-import * as ts from "typescript";
 import {
     ContainerReflection,
     DeclarationReflection,
     makeRecursiveVisitor,
     ParameterReflection,
     ProjectReflection,
+    ReferenceType,
     Reflection,
     SignatureReflection,
     TypeParameterReflection,
 } from "../models";
 import type { Logger } from "../utils";
+import { nicePath } from "../utils/paths";
 
 function makeIntentionallyExportedHelper(
+    project: ProjectReflection,
     intentional: readonly string[],
     logger: Logger
 ) {
@@ -25,7 +27,7 @@ function makeIntentionallyExportedHelper(
     });
 
     return {
-        has(symbol: ts.Symbol, typeName: string) {
+        has(type: ReferenceType, typeName: string) {
             // If it isn't declared anywhere, we can't produce a good error message about where
             // the non-exported symbol is, so even if it isn't ignored, pretend it is. In practice,
             // this will happen incredibly rarely, since symbols without declarations are very rare.
@@ -34,29 +36,20 @@ function makeIntentionallyExportedHelper(
             // 2. Properties on non-homomorphic mapped types, e.g. the symbol for "foo" on `Record<"foo", 1>`
             // There might be others, so still check this here rather than asserting, but print a debug log
             // so that we can possibly improve this in the future.
-            if (!symbol.declarations?.length) {
+            if (!type.package) {
                 logger.verbose(
-                    `The symbol ${symbol.name} has no declarations, implicitly allowing missing export.`
+                    `The type ${type.qualifiedName} has no declarations, implicitly allowing missing export.`
                 );
                 return true;
             }
 
             // Don't produce warnings for third-party symbols.
-            if (
-                symbol.declarations.some((d) =>
-                    d.getSourceFile().fileName.includes("node_modules")
-                )
-            ) {
+            if (type.package !== project.packageName) {
                 return true;
             }
 
             for (const [index, [file, name]] of processed.entries()) {
-                if (
-                    typeName === name &&
-                    symbol.declarations.some((d) =>
-                        d.getSourceFile().fileName.endsWith(file)
-                    )
-                ) {
+                if (typeName === name && type.sourceFileName?.endsWith(file)) {
                     used.add(index);
                     return true;
                 }
@@ -78,31 +71,30 @@ export function validateExports(
     let current: Reflection | undefined = project;
     const queue: Reflection[] = [];
     const intentional = makeIntentionallyExportedHelper(
+        project,
         intentionallyNotExported,
         logger
     );
-    const warned = new Set<ts.Symbol>();
+    const warned = new Set<string>();
 
     const visitor = makeRecursiveVisitor({
         reference(type) {
-            // If we don't have a symbol, then this was an intentionally broken reference.
-            const symbol = type.getSymbol();
-            if (!type.reflection && symbol) {
-                if (
-                    (symbol.flags & ts.SymbolFlags.TypeParameter) === 0 &&
-                    !intentional.has(symbol, type.qualifiedName) &&
-                    !warned.has(symbol)
-                ) {
-                    warned.add(symbol);
-                    const decl = symbol.declarations![0];
+            const uniqueId = `${type.sourceFileName}\0${type.qualifiedName}`;
 
-                    logger.warn(
-                        `${
-                            type.qualifiedName
-                        } is referenced by ${current!.getFullName()} but not included in the documentation.`,
-                        decl
-                    );
-                }
+            // If we don't have a symbol, then this was an intentionally broken reference.
+            if (
+                !type.reflection &&
+                !type.isIntentionallyBroken() &&
+                !intentional.has(type, type.qualifiedName) &&
+                !warned.has(uniqueId)
+            ) {
+                warned.add(uniqueId);
+
+                logger.warn(
+                    `${type.qualifiedName}, defined in ${nicePath(
+                        type.sourceFileName!
+                    )}, is referenced by ${current!.getFullName()} but not included in the documentation.`
+                );
             }
         },
         reflection(type) {